Refactor: Load default model and update readme
This commit addresses issues with loading the default model and
improves compatibility with older PowerShell versions.

The changes were motivated by the need to support older PowerShell
versions and ensure the default model is loaded correctly. The null
coalescing operator was replaced with an `if/else` statement which is
backwards compatible.

*   Replaced the null coalescing operator (`??`) with an `if/else`
    statement for loading the default model in both
    `Invoke-GeminiAIGenAPI.ps1` and `Invoke-OllamaAIGenAPI.ps1`. This
    enhances compatibility with older PowerShell versions.
*   Updated the readme.md to reflect support for PowerShell 5.1 and
    higher, removing the previous requirement for PowerShell Core 7.0
    or higher.